refactor(input/images): extract frame-reader helpers into reusable header
Move discover_frames, has_image_extension and natural_less out of
test/src/proxyncorr.cpp into a new header-only unit
include/ncorr/frame_reader.h (namespace ncorr, inline functions) so they
can be unit-tested in isolation without compiling the driver. Behaviour
is unchanged; proxyncorr.cpp now includes the header and uses the
functions via its existing 'using namespace ncorr'.
